pyproject.toml
[project]
name = "replm"
version = "0.1.0"
description = "Recursive Language Models — process arbitrarily long prompts by offloading context into a REPL and enabling symbolic recursion via sub-LLM calls"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "replm contributors" },
]
keywords = ["llm", "recursive", "long-context", "repl", "openai", "inference"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "openai>=1.0.0",
]

[project.optional-dependencies]
examples = [
    "jupyter>=1.0",
    "python-dotenv>=1.0",
]
benchmarks = [
    "matplotlib>=3.8",
    "python-dotenv>=1.0",
]
tracing = [
    "opentelemetry-api>=1.20",
]

[project.urls]
Homepage = "https://github.com/dschulmeist/replm"
Documentation = "https://github.com/dschulmeist/replm#readme"
Repository = "https://github.com/dschulmeist/replm"
Issues = "https://github.com/dschulmeist/replm/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/replm"]

[dependency-groups]
dev = [
    "mypy>=1.10",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "python-dotenv>=1.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"benchmarks/s_niah/filler.py" = ["E501"]
"benchmarks/aggregation/questions.py" = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
